[EE] Fix crashes upon entering game

+++ common/src/Engine.c
@@ -34,10 +34,10 @@ enum EGameMenuState nextNavigationSelection = kResumeCurrentState;
 long timeUntilNextState = MENU_ITEM_TIME_TO_BECOME_ACTIVE_MS;
 enum EPresentationState currentPresentationState;
 size_t biggestOption;
-uint8_t isRunning = TRUE;
+int isRunning = TRUE;
 const char *mainText = NULL;
-uint8_t enable3DRendering = TRUE;
-uint8_t needsToRedrawHUD = 1;
+int enable3DRendering = TRUE;
+int needsToRedrawHUD = 1;
 
 void enterState(enum EGameMenuState newState) {
